LPS-61786 Move layout finders into "portlet" package
